pyproject.toml
[project]
name = "intentlayer-aiserver"
version = "0.1.0"
description = "Serveur IA modulaire pour analyse NLP et génération d'UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "openai>=1.3.0",
    "langchain>=0.0.350",
    "langchain-openai>=0.0.2",
    "spacy>=3.7.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "langchain-community>=0.3.24",
    "nltk>=3.9.1",
    "pip>=25.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0"
]
